Fix integration tests, test event hook
